Inline similar.ts helper to fix Vercel ERR_MODULE_NOT_FOUND
In production, /api/similar crashed at cold-start with:

  Error [ERR_MODULE_NOT_FOUND]: Cannot find module
  '/var/task/api/_lib/similar' imported from /var/task/api/similar.js

Vercel compiles each api/*.ts file to .js individually but does not
rewrite relative imports. The compiled output retained
`import {...} from './_lib/similar'` (no extension), which Node ESM
rejects at runtime — even though both files were deployed.

Inline the Last.fm fetch + shaping helpers directly into api/similar.ts
to eliminate the relative import. The Vite dev middleware in
vite.config.ts now imports the same exports from api/similar.ts; the
`import type` from @vercel/node is erased at compile time and doesn't
leak into the dev bundle.
